add condition on tests
